Split the invite/scan screen into two dedicated sheets
One view showed a conversation's code and a camera behind a segmented
control, and it was presented as a full-screen overlay drawn inside
whatever surface opened it. The overlay carried its own back chevron,
title chip and add-people button, so it landed on top of the presenting
sheet's close button and conversation pill: three navigation affordances
for one screen, no composer, nothing dismissible in an obvious way.

There are two jobs here, and they are not the same job. Showing your
code is about this conversation. Scanning reads someone else's code and
opens whatever is behind it - it never adds anyone to the conversation
you started from. So each gets its own screen, presented as a sheet with
one chrome:

- `InviteCodeSheet` (the new `InviteCodeCard`): the QR, "Share invite
  link", and the line explaining what happens. Presented by the invite
  menu, conversation info, and the picker's "Show an invite code" row.
- `JoinConversationView`: the scanner, which already existed as a
  complete screen with its own close, paste, and invalid-code alert.
  The home's scan button now presents it directly instead of claiming a
  conversation just to host a viewfinder inside one; whatever it decodes
  is routed like any other link.

That removes the segmented control, the overlay, the transcript cell
branch that hosted the same body inline, and the `showsEmbeddedInvite` /
`showsInviteScanCard` plumbing threaded from the view models down
through the collection view's data source to the cell. The `.invite`
cell keeps its other half - the inviter's QR header and invite menu.
