+++ Convos/Conversation Detail/ConversationInfoView.swift
@@ -97,7 +97,8 @@ struct ConversationInfoView: View {
     @State private var presentingEditView: Bool = false
     @State private var showingLockedInfo: Bool = false
     @State private var showingFullInfo: Bool = false
-    @State private var presentingShareView: Bool = false
+    /// Drives the invite-code sheet the "Convo code" row presents.
+    @State private var presentingInviteCode: Bool = false
     /// Drives the system share sheet behind the menu's "Invite friends" row.
     @State private var presentingInviteShareSheet: Bool = false
 
@@ -109,11 +110,6 @@ struct ConversationInfoView: View {
         return [invite.inviteURLString]
     }
     @State private var presentingAddFromContactsPicker: Bool = false
-    /// Invite code scanned from the in-sheet Scan/Invite overlay, parked until
-    /// this sheet finishes dismissing so the join sheet (presented by
-    /// `ConversationView` beneath) isn't dropped mid-dismissal. Delivered by
-    /// `onDisappear` via `deliverPendingScannedCodeIfNeeded`.
-    @State private var pendingScannedCode: String?
     @State private var exportedLogsURL: URL?
     @State private var metadataDebugText: String = "Loading…"
     @State private var showingRestoreInviteTagAlert: Bool = false
@@ -398,36 +394,6 @@ struct ConversationInfoView: View {
         }
     }
 
-    /// Routes a code scanned from the in-sheet Scan/Invite overlay. The
-    /// overlay is shown on this view's local `@State`, but the view model
-    /// handler only flips its own flag, so the local overlay is dismissed here
-    /// too. A regular invite code then opens the join flow via
-    /// `presentingNewConversationForInvite`, whose sheet hangs off
-    /// `ConversationView` beneath this sheet and cannot present while the info
-    /// sheet is up (setting it mid-dismissal gets dropped by SwiftUI too) --
-    /// so the code is parked in `pendingScannedCode`, the info sheet is
-    /// dismissed, and `onDisappear` hands the code over once the dismissal has
-    /// fully settled. Agent-template codes join in place with no
-    /// presentation, so the info sheet stays.
-    private func handleOverlayScannedCode(_ code: String) {
-        presentingShareView = false
-        let isAgentTemplate = URL(string: code).flatMap(DeepLinkHandler.agentTemplateId(from:)) != nil
-        guard !isAgentTemplate else {
-            viewModel.handleScannedCodeInCurrentConversation(code)
-            return
-        }
-        pendingScannedCode = code
-        dismiss()
-    }
-
-    /// Hands a parked scanned code to the view model once this sheet has left
-    /// the hierarchy, so the resulting join sheet presents cleanly.
-    private func deliverPendingScannedCodeIfNeeded() {
-        guard let code = pendingScannedCode else { return }
-        pendingScannedCode = nil
-        viewModel.handleScannedCodeInCurrentConversation(code)
-    }
-
     /// Opens the agent builder from this sheet's own `.sheet(item:)` so it
     /// stacks on top -- the chat view's builder sheet
     /// (`viewModel.presentAgentBuilder()`) would present beneath this
@@ -441,32 +407,11 @@ struct ConversationInfoView: View {
         }
     }
 
-    /// Local share-overlay presentation binding. Resets the requested initial
-    /// segment on dismissal (mirroring `ConversationPresenter`'s binding) so
-    /// the next plain convo-code open lands on the Invite tab, not a stale
-    /// Scan request from the contacts picker.
-    private var localShareOverlayBinding: Binding<Bool> {
-        Binding(
-            get: { presentingShareView },
-            set: { newValue in
-                presentingShareView = newValue
-                if !newValue {
-                    viewModel.shareViewInitialSegment = .invite
-                }
-            }
-        )
-    }
-
     var body: some View {
         infoContent
             .addFromContactsPicker(
                 viewModel: viewModel,
-                isPresented: $presentingAddFromContactsPicker,
-                // This view is itself a presented sheet: the presenter-level
-                // overlay (viewModel.presentingShareView) would open beneath
-                // it, so route the picker's Show-invite-code / Scan rows to
-                // the local in-sheet overlay instead.
-                onPresentShareOverlay: { presentingShareView = true }
+                isPresented: $presentingAddFromContactsPicker
             )
             .shareSheet(
                 isPresented: $presentingInviteShareSheet,
@@ -478,7 +423,6 @@ struct ConversationInfoView: View {
             }
             .onDisappear {
                 navigator?.closed(context: navState.closeContext())
-                deliverPendingScannedCodeIfNeeded()
             }
             .onChange(of: presentingEditView) { oldValue, newValue in
                 handleEditViewChanged(from: oldValue, to: newValue)
@@ -714,24 +658,11 @@ struct ConversationInfoView: View {
                     AgentsInfoView(onMakeAgent: { pendingAgentBuilderAfterIntro = true })
                         .padding(.top, 20)
                 })
-                .overlay {
-                    if presentingShareView {
-                        ConversationShareOverlay(
-                            conversation: viewModel.conversation,
-                            invite: viewModel.invite,
-                            isPresented: localShareOverlayBinding,
-                            topSafeAreaInset: 0,
-                            coreActions: viewModel.coreActions,
-                            // Honors the segment the contacts picker's rows
-                            // request (Show invite code -> .invite, Scan ->
-                            // .scan); the toolbar convo-code path leaves the
-                            // view model default (.invite) untouched.
-                            initialSegment: viewModel.shareViewInitialSegment,
-                            onScannedCode: { code in
-                                handleOverlayScannedCode(code)
-                            }
-                        )
-                    }
+                .sheet(isPresented: $presentingInviteCode) {
+                    InviteCodeSheet(
+                        conversation: viewModel.conversation,
+                        invite: viewModel.invite
+                    )
                 }
                 .background {
                     Color.clear
