+++ Convos/Conversation Detail/Messages/MessagesView.swift
@@ -105,14 +105,6 @@ struct MessagesView<BottomBarContent: View>: View {
     /// in production); the testtube button stays hidden in any other case.
     var onDebugAttachmentTap: (() -> Void)?
     var extraBottomInset: CGFloat = 0.0
-    /// When true the index-0 `.invite` cell renders the full inline
-    /// Invite/Scan card (`InviteCodeBody`) for an active hosted session.
-    /// Mirrors `ConversationView.showsTopOfConvoInvite`.
-    var showsInviteScanCard: Bool = false
-    var inviteScanMode: InviteCodeMode = .inConvo
-    var inviteScanInitialSegment: ScanInviteSegment = .invite
-    var onScannedInviteCode: ((String) -> Void)?
-    var onInviteShareCompleted: ((UIActivity.ActivityType?, Bool, Error?) -> Void)?
     /// False when the composer is hosted externally (the conversation
     /// sheet): the transcript renders no bar of its own and insets purely by
     /// `extraBottomInset`, which the sheet keeps fed with its measured
@@ -238,11 +230,6 @@ struct MessagesView<BottomBarContent: View>: View {
                     sentAt: sentAt
                 )
             },
-            showsInviteScanCard: showsInviteScanCard,
-            inviteScanMode: inviteScanMode,
-            inviteScanInitialSegment: inviteScanInitialSegment,
-            onScannedInviteCode: onScannedInviteCode,
-            onInviteShareCompleted: onInviteShareCompleted,
             agentBuilderSummaryProvider: representableAgentBuilderSummaryProvider,
             currentUserProfileImage: { ProfileSettingsViewModel.shared.profileImage },
             backwardsSecrecyInfoSheet: { AnyView(BackwardsSecrecyInfoView()) },
