+++ Convos/Shared Views/ComposerMovedTypes.swift
@@ -44,9 +44,6 @@ typealias CloudConnectionServiceCatalog = ConvosComposer.CloudConnectionServiceC
 typealias AttachmentShareLink = ConvosComposer.AttachmentShareLink
 typealias ConnectionServiceIcon = ConvosComposer.ConnectionServiceIcon
 typealias InviteCodeMode = ConvosComposer.InviteCodeMode
-typealias InviteCodeBody = ConvosComposer.InviteCodeBody
-typealias ScanInviteSegment = ConvosComposer.ScanInviteSegment
-typealias ScanInviteToggle = ConvosComposer.ScanInviteToggle
 typealias QRScannerView = ConvosComposer.QRScannerView
 typealias QRScannerViewModel = ConvosComposer.QRScannerViewModel
 typealias AgentActivatingCardView = ConvosComposer.AgentActivatingCardView
