+++ ConvosCore/Sources/ConvosComposer/Messages View Controller/View Controller/Cells/MessagesListItemTypeCell.swift
@@ -87,25 +87,17 @@ class MessagesListItemTypeCell: UICollectionViewCell {
                     MessagesListItemTypeCell.messagesGroupContent(group: group, config: config)
 
                 case .invite(let invite):
-                    if config.showsInviteScanCard, let conversation = config.inviteScanConversation {
-                        MessagesListItemTypeCell.inviteScanCardContent(
-                            invite: invite,
-                            conversation: conversation,
-                            config: config
-                        )
-                    } else {
-                        VStack(spacing: DesignConstants.Spacing.step4x) {
-                            if config.headerMode == .standard, !config.hidesInviteCard {
-                                InviteView(invite: invite)
-                            }
-                            NewConvoIdentityView(
-                                onCopyLink: config.onCopyInviteLink,
-                                onConvoCode: config.onConvoCode
-                            )
+                    VStack(spacing: DesignConstants.Spacing.step4x) {
+                        if config.headerMode == .standard, !config.hidesInviteCard {
+                            InviteView(invite: invite)
                         }
-                        .padding(.vertical, DesignConstants.Spacing.step4x)
-                        .padding(.horizontal, DesignConstants.Spacing.step4x)
+                        NewConvoIdentityView(
+                            onCopyLink: config.onCopyInviteLink,
+                            onConvoCode: config.onConvoCode
+                        )
                     }
+                    .padding(.vertical, DesignConstants.Spacing.step4x)
+                    .padding(.horizontal, DesignConstants.Spacing.step4x)
 
                 case .conversationInfo(let conversation):
                     VStack(spacing: DesignConstants.Spacing.step4x) {
@@ -212,25 +204,6 @@ class MessagesListItemTypeCell: UICollectionViewCell {
     /// `InviteCodeBody` the full-screen `InviteCodeOverlay` composes, so the
     /// toggle + tabs don't fork. Extracted to keep the `setup` switch body
     /// under the type-check budget.
-    @ViewBuilder
-    fileprivate static func inviteScanCardContent(
-        invite: Invite,
-        conversation: Conversation,
-        config: CellConfig
-    ) -> some View {
-        InviteCodeBody(
-            conversation: conversation,
-            encodedURLString: invite.inviteURLString,
-            mode: config.inviteScanMode,
-            initialSegment: config.inviteScanInitialSegment,
-            isInviteReady: !invite.isEmpty,
-            onScannedCode: config.onScannedInviteCode,
-            onShareCompleted: config.onInviteShareCompleted
-        )
-        .padding(.vertical, DesignConstants.Spacing.step4x)
-        .frame(maxWidth: .infinity)
-    }
-
     @ViewBuilder
     fileprivate static func messagesGroupContent(group: MessagesGroup, config: CellConfig) -> some View {
         MessagesGroupView(
