+++ ConvosCore/Sources/ConvosComposer/Messages View Controller/View Controller/Data Source/CellFactory.swift
@@ -93,17 +93,6 @@ struct CellConfig {
     /// from the conversation-info preview. The view lives in the app
     /// target; when nil the explainer button doesn't present anything.
     let backwardsSecrecyInfoSheet: (() -> AnyView)?
-    /// When true the `.invite` cell renders the full inline Invite/Scan card
-    /// (`InviteCodeBody`) instead of the regular inviter QR + menu. Mirrors
-    /// `ConversationView.showsTopOfConvoInvite`.
-    let showsInviteScanCard: Bool
-    /// The conversation the inline Invite/Scan card renders for. Nil when the
-    /// card is not shown; the `.invite` cell guards on it.
-    let inviteScanConversation: Conversation?
-    let inviteScanMode: InviteCodeMode
-    let inviteScanInitialSegment: ScanInviteSegment
-    let onScannedInviteCode: (String) -> Void
-    let onInviteShareCompleted: (UIActivity.ActivityType?, Bool, Error?) -> Void
 }
 
 // swiftlint:disable force_cast
